Added route files for listing, index, and create_listing pages. Also added starter code to __init__.py and app.py.
