+++ .gitignore
@@ -1,6 +1,4 @@
 venv
 __pycache__
 .vscode
-<<<<<<< HEAD
-.DS_Store
 .DS_Store
