+++ app/routes/main_routes.py
@@ -1,0 +1,11 @@
+from flask import Blueprint, render_template, request, redirect, url_for
+
+main = Blueprint('main', __name__)
+
+@main.route('/')
+def index():
+    return render_template('index.html')
+
+@main.route('/profile', methods=['POST','GET'])
+def profile():
+    return render_template('profile.html')
